Remove unwanted intermediate function
As the function passed to render property only just takes props and passes it to the rendered component ( Cat ), We can pass a reference to the rendered component ( Cat ) directly then destructure the passed data in it and rename data with whatever we need. As a result, we will not have the issue of using the Pure component because we are passing the same reference, so we don't have to put a Caveats section.